backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
pydantic-settings>=2.1.0
httpx>=0.26.0
pytesseract>=0.3.10
pillow>=10.0.0
pypdf>=4.0.0
numpy>=1.26.0
opencv-python-headless>=4.8.0
pytest>=7.4.0
transformers>=4.36.0
torch>=2.1.0
reportlab>=4.0.0